fix rerendering and fetch data
