+++ src/component/missions/Missions.js
@@ -1,16 +1,11 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { getMissions, joinMission } from '../redux/missions/missionsSlice';
+import { joinMission } from '../redux/missions/missionsSlice';
 import './mission.css';
 
 export default function Missions() {
   const dispatch = useDispatch();
   const isLoading = useSelector((state) => state.missions.isLoading);
-
-  useEffect(() => {
-    dispatch(getMissions());
-  }, [dispatch]);
 
   const mission = useSelector((state) => state.missions.missions);
 
@@ -41,18 +36,19 @@ export default function Missions() {
             <tr>
               <td>{item.mission_name}</td>
               <td>{item.description}</td>
+              <td>{item.joined ? 'member' : 'Not a member'}</td>
               <td>
-                {item.joined ? 'member' : 'Not a member'}
-              </td>
-              <td>
-                <button type="button" onClick={() => dispatch(joinMission(item.mission_id))}>
-                  {item.joined ? 'leave Mission' : 'join Mission' }
+                <button
+                  className="joinMission"
+                  type="button"
+                  onClick={() => dispatch(joinMission(item.mission_id))}
+                >
+                  {item.joined ? 'leave Mission' : 'join Mission'}
                 </button>
               </td>
             </tr>
           </tbody>
         ))}
-
       </table>
     </div>
   );
